requirements.txt
youtube-transcript-api
openai
python-dotenv
streamlit
yt-dlp
openai-whisper
